added a function isEqual.ts
