+++ src/hocs/withRouter.ts
@@ -1,9 +1,7 @@
-import Block from '../utils/Block';
+import { Props } from '../utils/Block';
 import Router from '../utils/Router';
 
-export function withRouter(Component: any){
-	// @ts-ignore
-	type Props = typeof Component extends typeof Block<infer P> ? P : any;
+export function withRouter(Component: any) {
 
 	return class WithRouter extends Component {
 		constructor(props: Props & PropsWithRouter) {
