+++ src/utils/Router.ts
@@ -1,47 +1,5 @@
 import Block from './Block';
-
-type PlainObject<T = any> = {
-	[k in string]: T;
-};
-
-function isPlainObject(value: unknown): value is PlainObject {
-	return typeof value === 'object'
-		&& value !== null
-		&& value.constructor === Object
-		&& Object.prototype.toString.call(value) === '[object Object]';
-}
-
-function isArray(value: unknown): value is [] {
-	return Array.isArray(value);
-}
-
-function isArrayOrObject(value: unknown): value is [] | PlainObject {
-	return isPlainObject(value) || isArray(value);
-}
-
-function isEqual(lhs: any, rhs: any) {
-	if(typeof lhs === 'string' && typeof rhs === 'string') return lhs === rhs;
-
-	if (Object.keys(lhs).length !== Object.keys(rhs).length) {
-		return false;
-	}
-
-	for (const [key, value] of Object.entries(lhs)) {
-		const rightValue = rhs[key];
-		if (isArrayOrObject(value) && isArrayOrObject(rightValue)) {
-			if (isEqual(value, rightValue)) {
-				continue;
-			}
-			return false;
-		}
-
-		if (value !== rightValue) {
-			return false;
-		}
-	}
-
-	return true;
-}
+import isEqual from './isEqual';
 
 function render(query: string, block: Block) {
 	const root = document.querySelector(query);
@@ -113,7 +71,7 @@ class Router {
 			const target = event.currentTarget as Window;
 
 			this._onRoute(target.location.pathname);
-		}
+		};
 
 		this._onRoute(window.location.pathname);
 	}
